Added health check endpoint /healthz Add permanent hosting for VibeStamp #1
